Finalized Car Logistics for Orders
Added a car logistics object to separate the pick up / drop off time and place from the car object, since it wasn't appropriate to have those attributes for the same object. Also implemented a modal that would display the logistics in the history page. I will flesh this out later and put all the details of the order in this modal for the user to view.